backend/authentication_service/requirements.txt
Django>=3.0,<4.0
gunicorn
psycopg2>=2.8
djangorestframework
